pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gcode-proxy"
version = "0.1.0"
description = "A GCode proxy server that forwards GCode commands between TCP clients and USB serial devices"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "GCode Proxy Authors" }
]
keywords = ["gcode", "proxy", "serial", "3d-printing", "cnc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: Utilities",
]
dependencies = [
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
gcode-proxy-server = "gcode_proxy.cli:main"

[project.urls]
Homepage = "https://github.com/example/gcode-proxy"
Repository = "https://github.com/example/gcode-proxy"

[tool.hatch.build.targets.wheel]
packages = ["src/gcode_proxy"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]